% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_dimensions}
\alias{check_dimensions}
\title{Checks the dimensions of A, B, Lambda and Psi.}
\usage{
check_dimensions(A, B, Lambda, Psi)
}
\description{
Checks the dimensions of A, B, Lambda and Psi.
}
