% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilities.R
\name{coefficient_H}
\alias{coefficient_H}
\title{Calculate coefficient H.}
\usage{
coefficient_H(object)
}
\arguments{
\item{object}{Either a lavaan object or a list containing the factor loadings
and the residual errror covariance matrix.}
}
\value{
Coefficient H.
}
\description{
Calculate coefficient H.
}
\keyword{internal}
