% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{is_diagonal}
\alias{is_diagonal}
\title{Checks if a matrix is diagonal.}
\usage{
is_diagonal(x, eps = .Machine$double.eps^1/2)
}
\arguments{
\item{x}{A matrix.}

\item{eps}{A tolerance level.}
}
\value{
\code{TRUE} if the matrix is diagonal, \code{FALSE} otherwise.
}
\description{
Checks if a matrix is diagonal.
}
\keyword{internal}
