% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{is_scalar}
\alias{is_scalar}
\title{Checks if argument is a scalar.}
\usage{
is_scalar(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is scalar, \code{FALSE} otherwise.
}
\description{
Checks if argument is a scalar.
}
\keyword{internal}
