% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lavaan_to_list}
\alias{lavaan_to_list}
\title{Extracts matrices from a lavaan object.}
\usage{
lavaan_to_list(object)
}
\arguments{
\item{object}{A \code{lavaan} object.}
}
\value{
A list containing Lambda, Psi, and Gamma.
}
\description{
Extracts matrices from a lavaan object.
}
