% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilities.R
\name{omega}
\alias{omega}
\title{Calculate the congeneric reliabiltiy.}
\usage{
omega(object)
}
\arguments{
\item{object}{Either a lavaan object or a list containing the factor loadings
and the residual errror covariance matrix.}
}
\value{
The generalized congeneric reliability.
}
\description{
Calculate the congeneric reliabiltiy.
}
\keyword{internal}
