% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{omega_}
\alias{omega_}
\title{Calculate the generalized congeneric reliabiltiy.}
\usage{
omega_(Lambda, Psi, Gamma, A = 1, B = 0)
}
\arguments{
\item{Lambda}{A positive definite matrix or vector.}

\item{Psi}{A positive definite diagonal matrix.}

\item{Gamma}{A positive definite matrix.}

\item{A}{A column vector.}

\item{B}{A matrix.}
}
\value{
The generalized congeneric reliability.
}
\description{
Calculate the generalized congeneric reliabiltiy.
}
\keyword{internal}
