% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rsq_linear}
\alias{rsq_linear}
\title{Calculate R squared for a the linear action space.}
\usage{
rsq_linear(Lambda, Psi, Gamma, A, B)
}
\arguments{
\item{Lambda}{A positive definite matrix.}

\item{Psi}{A positive definite matrix.}

\item{A}{A matrix.}

\item{B}{A matrix.}
}
\value{
The generalized congeneric reliability.
}
\description{
Calculate R squared for a the linear action space.
}
\keyword{internal}
