% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rsq_sumscore}
\alias{rsq_sumscore}
\title{Calculate R squared for a the sum score action space}
\usage{
rsq_sumscore(Lambda, Psi, Gamma, A, B, iota)
}
\arguments{
\item{Lambda}{A positive definite matrix.}

\item{Psi}{A positive definite matrix.}

\item{A}{A matrix.}

\item{B}{A matrix.}

\item{iota}{A matrix of \code{-1, 0, 1}.}
}
\value{
The generalized congeneric reliability.
}
\description{
Calculate R squared for a the sum score action space
}
\keyword{internal}
