% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{to_list}
\alias{to_list}
\title{Extract matrices from a lavaan object or return a list.}
\usage{
to_list(object)
}
\arguments{
\item{object}{A \code{lavaan} object or a list of Lambda, Psi and Gamma.}
}
\value{
A list containing Lambda, Psi, and Gamma.
}
\description{
Extract matrices from a lavaan object or return a list.
}
\keyword{internal}
