% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tr}
\alias{tr}
\title{Calculate the trace of a matrix.}
\usage{
tr(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
The trace of the matrix.
}
\description{
Calculate the trace of a matrix.
}
\keyword{internal}
