% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{greedy}
\alias{greedy}
\title{Greedy algorithm for finding the optimal iota vector}
\usage{
greedy(A, B)
}
\arguments{
\item{A}{a positive definite matrix.}

\item{B}{a positive definite matrix with the same dimension as A.}
}
\description{
Greedy algorithm for finding the optimal iota vector
}
\keyword{internal}
